+++ MTPHandler/examples/__init__.py
@@ -1,2 +1,2 @@
-from .spectramax_parser import parse_spectramax
-from .megellan_parser import parse_magellan
+from ..ioutils.spectramax_parser import parse_spectramax
+from ..ioutils.megellan_parser import parse_magellan
